refactor canvas
