+++ ml4a/__init__.py
@@ -1,7 +1,6 @@
-from . import canvas
-from . import mask
-from . import video
-from . import util
+from .canvas import canvas
+from .canvas import mask
+from . import image
 
 __author__ = 'ml4a.net'
 __email__ = 'gene@genekogan.com'
