--- ml4a/models/__init__.py
@@ -1,4 +1,0 @@
-print("import ml4a/models/__init__")
-
-#from . import stylegan2
-
